+++ app/controllers/folio/console/searches_controller.rb
@@ -4,6 +4,8 @@ class Folio::Console::SearchesController < Folio::Console::BaseController
   add_breadcrumb I18n.t("folio.console.breadcrumbs.searches"), :console_search_path
 
   def show
+    authorize! :multisearch_console, current_site
+
     @query = ActionController::Base.helpers.sanitize(params[:q].to_s)
 
     respond_to do |format|