+++ config/initializers/audited.rb
@@ -2,3 +2,4 @@
 
 # when in impersonation, still use real user
 Audited.current_user_method = :true_user
+Audited.auditing_enabled = !(Rails.env.test?)