+++ test/cells/folio/console/layout/audited_dropdown_cell_test.rb
@@ -4,17 +4,19 @@ require "test_helper"
 
 class Folio::Console::Layout::AuditedDropdownCellTest < Folio::Console::CellTest
   test "show" do
-    page = create(:folio_page)
+    Audited.stub(:auditing_enabled, true) do
+      page = create(:folio_page)
 
-    html = cell("folio/console/layout/audited_dropdown",
-                page.revisions.reverse).(:show)
-    assert_not html.has_css?(".f-c-layout-audited-dropdown")
+      html = cell("folio/console/layout/audited_dropdown",
+                  page.revisions.reverse).(:show)
+      assert_not html.has_css?(".f-c-layout-audited-dropdown")
 
-    page.update!(title: "foo")
-    page.update!(title: "bar")
+      page.update!(title: "foo")
+      page.update!(title: "bar")
 
-    html = cell("folio/console/layout/audited_dropdown",
-                page.revisions.reverse).(:show)
-    assert html.has_css?(".f-c-layout-audited-dropdown")
+      html = cell("folio/console/layout/audited_dropdown",
+                  page.revisions.reverse).(:show)
+      assert html.has_css?(".f-c-layout-audited-dropdown")
+    end
   end
 end