+++ test/dummy/config/application.rb
@@ -10,13 +10,13 @@ require "folio"
 module Dummy
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
-    config.load_defaults 6.1
+    config.load_defaults 7.0
 
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    Rails.application.config.folio_leads = true
+    Rails.application.config.folio_leads_from_component_class_name = "Folio::Leads::FormComponent"
     Rails.application.config.folio_newsletter_subscriptions = true