+++ test/dummy/test/components/dummy/ui/documents_component_test.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require "test_helper"
+
+class Dummy::Ui::DocumentsComponentTest < Folio::ComponentTest
+  def test_render
+    model = "hello"
+
+    render_inline(Dummy::Ui::DocumentsComponent.new(model:))
+
+    assert_selector(".d-ui-documents")
+  end
+end